Remove RPC tests
